% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_tags.R
\name{collapse_tags}
\alias{collapse_tags}
\title{Collapse \code{term_count} Tags}
\usage{
collapse_tags(x, mapping, ...)
}
\arguments{
\item{x}{A \code{term_count} object.}

\item{mapping}{A list of named vectors where the vector names are the collapsed
column names and the vectors are the names of the columns to collapse.  Setting
a column name to \code{NULL} deletes these columns from the output.}

\item{\ldots}{ignored.}
}
\value{
Returns a \code{term_count} object.
}
\description{
Collapse (sum) tags/columns of a \code{term_count} object or remove columns
without changing termco class.
}

