% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colo.R
\name{colo}
\alias{colo}
\title{Make Regex to Locate Strings Containing Co-ocuring Substrings}
\usage{
colo(..., not = NULL)
}
\arguments{
\item{not}{A substring to exclude from consideration.}

\item{\ldots}{Terms that cooccur/collocate}
}
\value{
Returns a regular expression.  If Windows attempts to copy to
clipboard as well.
}
\description{
Make a regex to locate strings that contain >= 2 substrings with optional
negation.
}
\examples{
colo('dog', 'cat')
}
\keyword{collocate}
\keyword{cooccur}

