% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/termcoLite-package.R
\docType{package}
\name{termcoLite}
\alias{package-termcoLite}
\alias{termco}
\alias{termcoLite}
\alias{termcoLite-package}
\title{Regular Expression Counts of Terms and Substrings}
\description{
A small suite of functions used to count terms and substrings in strings.
}

