% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_names.R
\name{update_names}
\alias{update_names}
\title{Rename a term_count Object's Term Columns}
\usage{
update_names(x, old, new)
}
\arguments{
\item{x}{A \code{term_count} object.}

\item{old}{A vector of the current names.}

\item{new}{A vector of new names corresponding to the order of \code{old} names.}
}
\value{
Returns a renamed \code{term_count} object.
}
\description{
Safely rename a \code{term_count} object's term columns and attributes.
}

